Rename strata-search to strata-intelligence and consolidate search architecture
Move search_types, tokenizer, InvertedIndex, and BM25LiteScorer into
strata-engine so primitives can index documents on write paths. Rename
the search crate to strata-intelligence and restructure it as a thin
orchestration layer (HybridSearch, Fuser, DatabaseSearchExt) that
re-exports moved types for backward compatibility.

Wire InvertedIndex into KV, JSON, Event, State, and Run primitive
write paths with zero overhead when disabled. Add Command::Search to
the executor for first-class cross-primitive search. Include
architecture documentation.
