+++ crates/engine/src/primitives/tokenizer.rs
@@ -13,7 +13,7 @@
 /// # Example
 ///
 /// ```
-/// use strata_search::tokenizer::tokenize;
+/// use strata_engine::primitives::tokenizer::tokenize;
 ///
 /// let tokens = tokenize("Hello, World!");
 /// assert_eq!(tokens, vec!["hello", "world"]);
@@ -31,7 +31,7 @@ pub fn tokenize(text: &str) -> Vec<String> {
 /// # Example
 ///
 /// ```
-/// use strata_search::tokenizer::tokenize_unique;
+/// use strata_engine::primitives::tokenizer::tokenize_unique;
 ///
 /// let tokens = tokenize_unique("test test TEST");
 /// assert_eq!(tokens, vec!["test"]);
