+++ crates/executor/src/tests/mod.rs
@@ -5,3 +5,4 @@ pub mod execute_many;
 pub mod parity;
 pub mod determinism;
 pub mod session;
+pub mod search;
